% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{blocks}
\alias{blocks}
\title{Retrieve block ids, descriptions, and elements}
\usage{
blocks(design, elements = FALSE)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
A data.table giving the \code{id} and \code{description} of each
  block.
}
\description{
Retrieve block ids, descriptions, and elements
}

