% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.r
\name{choices}
\alias{choices}
\title{Get item response choices}
\usage{
choices(id)
}
\arguments{
\item{id}{A Qualtrics survey identifier}
}
\value{
A data.table of response choices
}
\description{
Request a table giving the response choices for each survey item.
}

