% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.r
\name{design}
\alias{design}
\title{Export a survey design}
\usage{
design(id)
}
\arguments{
\item{id}{A Qualtrics survey identifier}
}
\value{
A list representing the survey design
}
\description{
Retrieve a JSON representation of a survey's design, as a list.
}

