% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sensitive.R
\name{drop_sensitive}
\alias{drop_sensitive}
\alias{drop_meta}
\alias{keep_questions}
\alias{drop_meta}
\alias{keep_questions}
\title{Drop columns from survey responses}
\usage{
drop_sensitive(tbl)

drop_meta(tbl)

keep_questions(tbl, design_object)
}
\arguments{
\item{tbl}{A data.frame of survey responses.}

\item{design_object}{A \code{\link{qualtrics_design-class}} object retrieved
from Qualtrics by \code{\link{design}}.}
}
\value{
The data.frame without the dropped columns.
}
\description{
\code{drop_sensitive} drops the columns that give approximate location, IP
address, contact name, and contact email from a table of survey responses.
\code{drop_meta} also drops other standard columns that don't represent
question responses. \code{keep_questions} is most strict, keeping only the
columns for survey questions; it drops columns with embedded data, for
example.
}
\details{
The following metadata columns are dropped by \code{drop_sensitive}, if they
exist in the table:
\itemize{
\item LocationLatitude
\item LocationLongitude
\item LocationAccuracy
\item IPAddress
\item RecipientLastName
\item RecipientFirstName
\item RecipientEmail
}

\code{drop_meta} drops those columns and any of the following:
\itemize{
\item ExternalDataReference
\item EndDate
\item Finished
\item ResponseID
\item ResponseSet
\item StartDate
\item Status
}
}
