% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.r
\name{edges}
\alias{edges}
\alias{nodes}
\title{Represent survey flows as node or edge tables}
\usage{
edges(design)

nodes(design, ids = FALSE)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\description{
Represent survey flows as node or edge tables
}

