% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_names.r
\name{export_names}
\alias{export_names}
\title{Get export names for survey response fields}
\usage{
export_names(design)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
A two-column data.table giving \code{id} and \code{export_name}
  pairs.
}
\description{
Retrieve the descriptive export names (e.g., \code{favorability}) associated
with question ids (e.g., \code{QID1}).
}
\details{
In some cases, there is more than one export name associated with a given
survey question identifier. Multiple-choice, multiple-answer questions have a
unique export name for each choice, but only one question identifier.
}
\examples{
\dontrun{ # requires API key
# set an API key and retrieve a survey design
key_from_file()
design = survey_design("SV_cuxfjYWRTB30ouh")
export_names(design)
}
}

