% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_id.R
\name{find_id}
\alias{find_id}
\title{Find survey ids by searching survey names}
\usage{
find_id(pattern, first = TRUE, ignore.case = TRUE, perl = TRUE,
  fixed = FALSE)
}
\arguments{
\item{pattern}{Character string containing a regular expression (or character
string for \code{fixed = TRUE}).}

\item{first}{Whether to return only the first match (default) or all matches.
See Details.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{fixed}{logical.  If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}
}
\value{
A named character vector giving the id(s) of matching surveys, or an
empty character vector if there are no matches.
}
\description{
\code{find_id} searches for survey names that match argument \code{pattern}.
All arguments but \code{first} are passed to \code{\link[base]{grepl}} for
pattern matching.
}
\details{
\code{find_id} retrieves the names and ids of all available surveys (using
\code{\link{surveys}}), and then uses \code{\link[base]{grepl}} to search the
names for matches.

For \code{first = TRUE} (the default), only the id of the first matching name
(if any) is returned. Matches are sorted by \code{lastModified} timestamp
(descending), so \code{first = TRUE} gives the id of the matching survey most
recently modified.
}
\seealso{
\code{\link{surveys}}
}

