% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.r
\name{get_api_key}
\alias{get_api_key}
\title{Get the Qualtrics API key}
\usage{
get_api_key()
}
\value{
A string giving an API key
}
\description{
The user is expected to save their Qualtrics API key either in a file named
\code{.api_key} in the working directory or define a global variable named
\code{api_key}. This function looks for the key in each place and returns it
if successful, or stops. Called by other package functions, but may also be
used directly to verify that the correct key will be used.
}

