% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.r
\name{get_responses}
\alias{get_responses}
\title{Export survey responses}
\usage{
get_responses(id = "SV_738wNt5GB4wrTaR", labels = TRUE)
}
\arguments{
\item{id}{A Qualtrics survey identifier}

\item{labels}{Whether to use question labels and answer codes}
}
\value{
A data.table of survey responses
}
\description{
Export survey responses
}

