% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_from_file.r
\name{key_from_file}
\alias{key_from_file}
\title{Set the environment variable for the API key from a file}
\usage{
key_from_file(file = ".api_key", verbose = FALSE)
}
\arguments{
\item{file}{A filename.}

\item{verbose}{Whether to print the result of reading from \code{file}, or
not (the default).}
}
\value{
Invisibly returns the value of the environment variable
  \code{QUALTRICS_KEY}.
}
\description{
The environment variable \code{QUALTRICS_KEY} should give an API key for
communication with the Qualtrics survey platform. \code{key_from_file} sets
it using the contents of a file.
}

