% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_flow}
\alias{plot_flow}
\title{Plot a survey flow}
\usage{
plot_flow(design_object)
}
\arguments{
\item{design_object}{A \code{\link{qualtrics_design-class}} object retrieved
from Qualtrics by \code{\link{design}}.}
}
\value{
\code{plot_flow} returns a visNetwork object.
}
\description{
\code{plot_flow} shows a survey's flow as a directed graph showing possible
pathways through the elements of the survey.
}
\seealso{
\code{\link{edges}}, \code{\link{nodes}}, \code{\link{render_flow}}
}

