% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{post}
\alias{post}
\title{Wrap POST requests}
\usage{
post(url = "https://az1.qualtrics.com/API/v3/", endpoint = NULL,
  body = NULL, ...)
}
\arguments{
\item{url}{The base API URL}

\item{endpoint}{The endpoint name, e.g. \code{users}}

\item{body}{A named list of body data to post}

\item{...}{Arguments passed to \link[httr]{content}}
}
\value{
The content of the request's response via \link[httr]{content}
}
\description{
Typically called by other functions, not directly. This function can be used
to access API endpoints without package functions. See the
\href{https://api.qualtrics.com}{API documentation}.
}

