% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{qget}
\alias{qget}
\title{Make GET requests}
\usage{
qget(endpoint = NULL, as, ...)
}
\arguments{
\item{endpoint}{An API endpoint, e.g., \code{"surveys"}}

\item{as}{Passed to \link{httr}[content]}

\item{...}{Further arguments to \link[httr]{GET} or \link[httr]{POST},
depending on \code{verb}}
}
\value{
The content of the response via \link[httr]{content}
}
\description{
Use this function to access API endpoints for which higher-level package
functions are not available.
}

