% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{qpost}
\alias{qpost}
\title{Make POST requests}
\usage{
qpost(endpoint = NULL, as, ...)
}
\arguments{
\item{endpoint}{An API endpoint, e.g., \code{"surveys"}}

\item{as}{Passed to \link{httr}[content]}

\item{...}{Further arguments to \link[httr]{GET} or \link[httr]{POST},
depending on \code{verb}}

\item{body}{A named list of body data to post}
}
\value{
The content of the request's response via \link[httr]{content}
}
\description{
Use this function to access API endpoints for which higher-level package
functions are not available.
}

