% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics-package.r
\docType{package}
\name{qualtrics}
\alias{qualtrics}
\alias{qualtrics-package}
\title{qualtrics}
\description{
This package aims to make working with Qualtrics surveys in R easier. Main
features:
}
\details{
\itemize{
  \item Download survey responses into an R session
  \item Download survey designs
  \item List survey metadata (e.g. active/closed status and expiration dates)
  \item Drop sensitive fields from survey responses
}

The package works by sending requests to the Qualtrics API. For
authentication, an API key is needed. The package looks for it first in the
current working directory in the file \code{.api_key}, and then as a global
variable \code{api_key}.

qualtrics focuses on survey responses and designs. To access API endpoints
that are not available through package functions, \code{\link{qget}} and
\code{\link{qpost}} request wrappers are available.

I'm aware of two other Qualtrics packages.
\href{https://github.com/jbryer/qualtrics}{Jason Bryer} wrote one in 2012.
\href{https://github.com/ericpgreen/qualtrics}{Eric Green} forked and revised
it for v2.3 of the Qualtrics API, most recently in 2014.
}

