% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{questions}
\alias{questions}
\title{Get survey questions}
\usage{
questions(design_object)
}
\arguments{
\item{design_object}{A \code{\link{qualtrics_design-class}} object retrieved
from Qualtrics by \code{\link{design}}.}
}
\value{
\code{questions} returns a table of question identifiers and text.
}
\description{
Retrieve identifiers and text for a given survey's questions.
}
\seealso{
\code{\link{blocks}}, \code{\link{choices}}
}
