% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_flow.R
\name{render_flow}
\alias{render_flow}
\title{Render a survey design with a Shiny app}
\usage{
render_flow(design_object)
}
\arguments{
\item{design_object}{A \code{\link{qualtrics_design-class}} object retrieved
from Qualtrics by \code{\link{design}}.}
}
\value{
\code{render_flow} invisibly returns the table underlying the Shiny
UI.
}
\description{
The Shiny app renders a visualization of the survey flow: blocks, branching,
randomization, etc. On selection of a block, its questions are visible.
Selecting a question in turn shows its text.
}
\examples{
\dontrun{
svy_design <- design("SV_0VVlb9QwJ4bsBKZ")
render_flow(svy_design)
}
}
\seealso{
\code{\link{plot_flow}} for a visNetwork plot, or
\code{\link{nodes}} and \code{\link{edges}} for node and edge data in tabular
form.
}
