% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_survey.r
\name{render_survey}
\alias{render_survey}
\title{Render a survey design with a Shiny app}
\usage{
render_survey(design)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
Invisibly returns the data.table underlying the Shiny UI.
}
\description{
Render a survey design with a Shiny app
}

