% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{request}
\alias{request}
\title{Make requests of the Qualtrics API}
\usage{
request(verb = "GET", subdomain = "az1", endpoint = NULL,
  key = Sys.getenv("QUALTRICS_KEY"), test = FALSE, ...)
}
\arguments{
\item{verb}{Name of verb to use.}

\item{subdomain}{A Qualtrics subdomain, e.g. \code{"ca1"}; see the
\href{https://api.qualtrics.com/docs/root-url}{Qualtrics documentation}}

\item{endpoint}{An API endpoint, e.g., \code{"surveys"}}

\item{key}{A Qualtrics API key}

\item{test}{Whether to send the request to \url{https://httpbin.org} for
testing purposes}

\item{...}{Further arguments to \link[httr]{GET} or \link[httr]{POST},
depending on \code{verb}}
}
\value{
A \link[httr]{response} object
}
\description{
Make requests of the Qualtrics API
}

