% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_count.r
\name{response_count}
\alias{response_count}
\title{Get response counts for a survey}
\usage{
response_count(id)
}
\arguments{
\item{id}{A Qualtrics survey identifier}
}
\value{
A list giving response counts
}
\description{
Response counts are given by category. \code{generated} responses are those
created by the "Test Survey" tool. \code{deleted} responses were removed by
the "Delete Data" tool.
}

