% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.r
\name{surveys}
\alias{surveys}
\title{Download survey metadata}
\usage{
surveys(key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"))
}
\arguments{
\item{key}{A Qualtrics API key (by default, the value of the environment
variable \code{QUALTRICS_KEY}).}

\item{subdomain}{A Qualtrics subdomain (by default, the value of the
environment variable \code{QUALTRICS_SUBDOMAIN}).}
}
\value{
A data.table of survey metadata
}
\description{
Retrieve names, identifiers, and other metadata for all available surveys.
The result is sorted by the \code{last_modified} timestamp in descending
order.
}
\examples{
surveys()
}

