% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_responses.r
\name{values_to_labels}
\alias{labels_to_values}
\alias{values_to_labels}
\title{Recode survey responses}
\usage{
values_to_labels(tbl, id)

labels_to_values(tbl, id)
}
\arguments{
\item{tbl}{A data.frame of survey responses.}

\item{id}{A Qualtrics survey identifier.}
}
\value{
The data.frame with recoded responses.
}
\description{
\code{values_to_labels} recodes survey responses that contain choice values
or ids (like \code{"5"}) using descriptive labels (like \code{"strongly
approve"}). \code{labels_to_values} does the reverse.
}

