% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{blocks}
\alias{blocks}
\title{Get block ids, descriptions, and elements}
\usage{
blocks(design, elements = FALSE)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
A data.table giving the \code{block_id} and \code{block_description}
of each block, and optionally its elements.
}
\description{
Retrieve each block's \code{id}, as generated by Qualtrics, and
\code{description}, as assigned by the user in the control panel. Optionally,
if argument \code{elements = TRUE}, \code{blocks} also shows the elements of
each block (questions and page breaks), in a long table of block-element
pairs.
}

