% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.r
\name{choices}
\alias{choices}
\title{Get survey response choices}
\usage{
choices(design)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
A data.table of response choices.
}
\description{
Retrieve a table giving the response choices for each question in a survey.
}
\seealso{
Download a survey's \code{\link{questions}},
  \code{\link{responses}}, or \code{\link{design}}.
}

