% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\alias{qualtrics_design-class}
\title{Download a survey's design}
\usage{
design(id, key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), verbose = FALSE)
}
\arguments{
\item{id}{A Qualtrics survey identifier.}

\item{key}{A Qualtrics API key (by default, the value of the environment
variable \code{QUALTRICS_KEY}).}

\item{subdomain}{A Qualtrics subdomain (by default, the value of the
environment variable \code{QUALTRICS_SUBDOMAIN}).}

\item{verbose}{Print progress.}
}
\value{
\code{design()} returns a \code{qualtrics_design} object, which is an
S3 class built from a list.
}
\description{
Retrieve a representation of the elements in a survey and their
configuration.
}
\examples{
\dontrun{
design_object <- design("SV_0VVlb9QwJ4bsBKZ")
print(design_object)
}
}
\seealso{
\code{\link{responses}}
}
