% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.r
\name{design}
\alias{design}
\title{Download a survey's design}
\usage{
design(id, as = "parsed")
}
\arguments{
\item{id}{A Qualtrics survey identifier.}

\item{as}{Desired type of content. See \link[httr]{content}.}
}
\value{
By default, a list representing the survey design; see \code{as}.
}
\description{
Retrieve a representation of the elements in a given survey.
}
\seealso{
Download a survey's \code{\link{responses}},
  \code{\link{questions}}, or question \code{\link{choices}}.
}

