% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sensitive.r
\name{drop_sensitive}
\alias{drop_sensitive}
\title{Drop sensitive columns from survey responses}
\usage{
drop_sensitive(tbl)
}
\arguments{
\item{tbl}{A data.frame of survey responses.}
}
\value{
The data.frame without the sensitive columns.
}
\description{
Drop the columns that give approximate location, IP address, contact name,
and contact email from a table of survey responses.
}
\details{
The following metadata columns are dropped, if they exist in the table:
\itemize{
\item LocationLatitude
\item LocationLongitude
\item LocationAccuracy
\item IPAddress
\item RecipientLastName
\item RecipientFirstName
\item RecipientEmail
}
}

