% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{edges}
\alias{edges}
\alias{nodes}
\title{Represent survey flows as node or edge tables}
\usage{
edges(design_object)

nodes(design_object)
}
\arguments{
\item{design_object}{A \code{\link{qualtrics_design-class}} object retrieved
from Qualtrics by \code{\link{design}}.}
}
\description{
A survey flow can usefully be considered a directed graph. \code{edges} shows
its edge pairs, and \code{nodes} gives its nodes.
}
\seealso{
\code{\link{plot_flow}} for a visNetwork graph of a survey flow.
}

