% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_names.R
\name{export_names}
\alias{export_names}
\title{Get export names for survey response fields}
\usage{
export_names(design_object)
}
\arguments{
\item{design_object}{A \code{\link{qualtrics_design-class}} object retrieved
from Qualtrics by \code{\link{design}}.}
}
\value{
A two-column giving \code{question_id} and \code{export_name}
  pairs. See details.
}
\description{
Retrieve the descriptive export names associated with question ids (e.g.,
\code{QID1}).
}
\details{
In some cases, there is more than one export name associated with a given
survey question identifier. Multiple-choice, multiple-answer questions have a
unique export name for each choice, but only one question identifier.
}
\seealso{
\code{\link{questions}}
}

