% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.r
\name{get_survey_design}
\alias{get_survey_design}
\title{Export a survey design}
\usage{
get_survey_design(survey_id = "SV_738wNt5GB4wrTaR")
}
\arguments{
\item{survey_id}{The survey identifier as a string}
}
\value{
A list representing the survey design
}
\description{
Export a survey design
}

