% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_questions.R
\name{ids_to_names}
\alias{ids_to_names}
\alias{names_to_ids}
\title{Rename response table columns}
\usage{
ids_to_names(tbl, design_object)

names_to_ids(tbl, design_object)
}
\arguments{
\item{tbl}{A data.frame of survey responses.}

\item{design_object}{A \code{\link{qualtrics_design-class}} object retrieved
from Qualtrics by \code{\link{design}}.}
}
\value{
The table with renamed columns.
}
\description{
\code{ids_to_names} takes a table of responses and renames the columns with
question identifier names (e.g. \code{QID2}) using question names (e.g.,
\code{approval}). \code{names_to_ids} does the reverse.
}

