% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.r
\name{plot_flow}
\alias{edges}
\alias{plot_flow}
\title{Represent a survey design as a graph}
\usage{
plot_flow(design)

edges(design)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
\code{plot_flow} invisibly returns a DiagrammeR \code{dgr_graph}
  object. As a side effect, it renders the graph. \code{edges} returns a
  data.table of graph edges.
}
\description{
\code{plot_flow} and \code{edges} represent a survey's design as a directed
graph that shows possible paths through the elements of the survey.
}

