% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_flow}
\alias{edges}
\alias{plot_flow}
\title{Represent a survey design as a graph}
\usage{
plot_flow(design)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
\code{plot_flow} invisibly returns a visNetwork object. As a side
  effect, it renders the graph.
}
\description{
\code{plot_flow} and \code{edges} represent a survey's design as a directed
graph that shows possible paths through the elements of the survey.
}
\seealso{
\code{nodes} and \code{edges} for node and edge data in tabular
  form.
}

