% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{qget}
\alias{qget}
\title{Wrap GET requests}
\usage{
qget(url = "https://az1.qualtrics.com/API/v3/", endpoint = NULL,
  as = "parsed", ...)
}
\arguments{
\item{url}{The base API URL}

\item{endpoint}{The endpoint name, e.g. \code{users}}

\item{as}{Passed to \link{httr}[content]}

\item{...}{Arguments passed to \link[httr]{GET}}
}
\value{
The content of the request's response via \link[httr]{content}
}
\description{
Typically called by other functions, not directly. This function can be used
to access API endpoints without package functions. See the
\href{https://api.qualtrics.com}{API documentation}.
}

