% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.r
\name{questions}
\alias{questions}
\title{Get survey questions}
\usage{
questions(design, labels = TRUE, text = TRUE, html = FALSE)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}

\item{labels}{Retrieve question labels (default), or don't.}

\item{text}{Retrieve question text (default), or don't.}
}
\value{
A data.table of question names, labels, and (optionally) text
}
\description{
Retrieve the names, labels, and text of a given survey's questions.
}
\seealso{
Retrieve a survey's question \code{\link{choices}} or
  \code{\link{responses}}.
}

