% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.r
\name{questions}
\alias{questions}
\title{Get question names, labels, and text}
\usage{
questions(id, text = TRUE)
}
\arguments{
\item{id}{A Qualtrics survey identifier}

\item{text}{Retrieve question text (default), or don't}
}
\value{
A data.table of question names, labels, and (optionally) text
}
\description{
Retrieve the names, labels, and text of a given survey's questions.
}

