% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.r
\name{questions}
\alias{questions}
\title{Download survey questions}
\usage{
questions(id, labels = TRUE, text = TRUE)
}
\arguments{
\item{id}{A Qualtrics survey identifier.}

\item{labels}{Retrieve question labels (default), or don't.}

\item{text}{Retrieve question text (default), or don't.}
}
\value{
A data.table of question names, labels, and (optionally) text
}
\description{
Retrieve the names, labels, and text of a given survey's questions.
}
\seealso{
Download a survey's question \code{\link{choices}},
  \code{\link{responses}}, or \code{\link{design}}.
}

