% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_survey.r
\name{render_survey}
\alias{render_survey}
\title{Render a survey design with a Shiny app}
\usage{
render_survey(design)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
\code{render_survey} invisibly returns the table underlying the Shiny
UI.
}
\description{
The Shiny app renders a visualization of the survey flow: blocks, branching,
randomization, etc. On selection of a block, its questions are visible.
Selecting a question in turn shows its text. An example is available online
through shinyapps.io (TODO).
}

