% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{request}
\alias{qget}
\alias{qpost}
\alias{request}
\title{Send GET and POST requests to the Qualtrics API}
\usage{
request(verb = "GET", subdomain = "az1", action = NULL,
  key = Sys.getenv("QUALTRICS_KEY"), test = FALSE, ...)

qget(action = NULL, as = "parsed", ...)

qpost(action = NULL, as = "parsed", ...)
}
\arguments{
\item{verb}{Name of verb to use.}

\item{subdomain}{A Qualtrics subdomain.}

\item{action}{An API action (like \code{"surveys"}).}

\item{key}{A Qualtrics API key (by default, the value of the environmental
variable \code{QUALTRICS_KEY}).}

\item{test}{Whether to send the request to \url{https://httpbin.org} for
testing purposes, rather than to the Qualtrics API.}

\item{...}{Further arguments to \code{\link[httr]{GET}} or
\code{\link[httr]{POST}}, depending on \code{verb}.}

\item{as}{Desired type of content. See \link[httr]{content}.}
}
\value{
For \code{request}, a \code{\link[httr]{response}} object. For
  \code{qget} and \code{qpost}, its content as extracted by
  \code{\link[httr]{content}}.
}
\description{
These are lower-level functions for direct access to API actions. They may be
useful in programming or when higher-level package functions are not
available. \code{qget} and \code{qpost} are wrappers for \code{request}.
}
\details{
For help specifying valid subdomains and API actions, see the
\href{https://api.qualtrics.com/docs/root-url}{Qualtrics documentation}.

If \code{test} is \code{TRUE}, the API key is removed, and then the request
is sent to the value of \code{paste0("https://httpbin.org/", tolower(verb)}.
}

