% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{request}
\alias{qget}
\alias{qpost}
\alias{request}
\title{Send a request to the Qualtrics API}
\usage{
request(verb = "GET", action = NULL, key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), api_url = NULL,
  verbose = FALSE, ...)

qget(action = NULL, as = "parsed", key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), ...)

qpost(action = NULL, as = "parsed", key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), ...)
}
\arguments{
\item{verb}{Name of verb to use.}

\item{action}{An API action (like \code{"surveys"}).}

\item{key}{A Qualtrics API key (by default, the value of the environment
variable \code{QUALTRICS_KEY}).}

\item{subdomain}{A Qualtrics subdomain (by default, the value of the
environment variable \code{QUALTRICS_SUBDOMAIN}).}

\item{api_url}{A complete API URL. If used, argument \code{action} and
\code{subdomain} are ignored.}

\item{verbose}{Output API calls to \code{\link[base]{stderr}}.}

\item{...}{Further arguments to \code{\link[httr]{GET}} or
\code{\link[httr]{POST}}, depending on \code{verb}.}

\item{as}{Desired type of content. See \link[httr]{content}.}
}
\value{
For \code{request}, a \code{\link[httr]{response}} object. For
  \code{qget} and \code{qpost}, its content as extracted by
  \code{\link[httr]{content}}.
}
\description{
These are lower-level functions for direct access to API actions. They may be
useful in programming or when higher-level package functions are not
available. \code{qget} and \code{qpost} are wrappers for \code{request}.
}
\details{
For help with subdomains, see the
\href{https://api.qualtrics.com/docs/root-url}{Qualtrics documentation}. Each
Qualtrics account is assigned a subdomain, and using another will work but
produce a warning.
}

