% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_count.r
\name{response_counts}
\alias{response_counts}
\title{Download survey response counts}
\usage{
response_counts(design)
}
\arguments{
\item{design}{A \code{\link{qualtrics_design-class}} object.}
}
\value{
A list giving response counts.
}
\description{
Response counts are given in three categories defined by the Qualtrics API:
auditable, generated, and deleted. See Details.
}
\details{
Categories:
\itemize{
\item \code{auditable}: typical responses
\item \code{generated}: responses created using the Qualtrics control panel's
"Test Survey" tool
\item \code{deleted}: responses removed using the Qualtrics control panel's
"Delete Data" tool
}
}

