% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.r
\name{responses}
\alias{responses}
\title{Download survey responses}
\usage{
responses(id, format = "json", use_labels = TRUE, verbose = TRUE,
  key = Sys.getenv("QUALTRICS_KEY"), ...)
}
\arguments{
\item{id}{A Qualtrics survey identifier.}

\item{use_labels}{Use question labels and choice descriptions (default),
instead of question and identifiers.}

\item{verbose}{Print progress.}

\item{key}{A Qualtrics API key (by default, the value of the environment
variable \code{QUALTRICS_KEY}).}

\item{...}{Additional parameters for the \code{responseexports} API.}
}
\value{
A data.table of survey responses.
}
\description{
Retrieve the responses to a survey.
}
\details{
\code{responses} requests that Qualtrics prepare a zipfile for download; once
it is ready, the archive is downloaded to a location given by
\code{\link[base]{tempfile}} and unzipped; its JSON contents read and parsed;
and the temporary files deleted.

By default, descriptive question labels and choice descriptions (e.g.,
\code{favorability} and \code{strongly approve}) are requested instead of
question ids and choice codes (e.g., \code{QID1} and {5}). This is controlled
by argument \code{use_labels}.

Other parameters that affect the format of the survey responses are
available. For example, passing \code{seenUnansweredRecode = "99"} in the
call to \code{responses} will show \code{"99"} as the response to unanswered
questions instead of \code{NA}. See the
\href{https://api.qualtrics.com/docs/json}{Qualtrics documentation} for the
details of these parameters. The parameters that cannot be changed via
\code{...} are \code{format}, \code{surveyId}, and \code{useLabels}, the last
of which is exposed as argument \code{use_labels}.
}
\seealso{
Retrieve a survey's \code{\link{questions}} or question
  \code{\link{choices}}.
}

