% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{responses}
\alias{responses}
\title{Download survey responses}
\usage{
responses(id, use_labels = TRUE, verbose = FALSE,
  key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), ...)
}
\arguments{
\item{id}{A Qualtrics survey identifier.}

\item{use_labels}{Use question labels and choice descriptions (default),
instead of question and identifiers.}

\item{verbose}{Print progress.}

\item{key}{A Qualtrics API key (by default, the value of the environment
variable \code{QUALTRICS_KEY}).}

\item{subdomain}{A Qualtrics subdomain (by default, the value of the
environment variable \code{QUALTRICS_SUBDOMAIN}).}

\item{...}{Additional parameters for the \code{responseexports} API.}
}
\value{
A table of survey responses.
}
\description{
Retrieve the responses to a survey.
}
\details{
Implementation: \code{responses} requests that Qualtrics prepare a zipfile
for download; once it is ready, the archive is downloaded to a location given
by \code{\link[base]{tempfile}} and unzipped; its JSON contents read and
parsed; and the temporary files deleted.

By default, descriptive question labels and choice descriptions are requested
instead of question ids and choice codes. This is controlled by argument
\code{use_labels}.

Other \href{https://api.qualtrics.com/docs/json}{documented} Qualtrics API
parameters that affect the format of survey responses are available. For
example, passing \code{seenUnansweredRecode = "99"} in the call to
\code{responses} will show \code{"99"} as the response to unanswered
questions instead of \code{NA}, the default. The parameters that cannot be
changed via \code{...} are \code{surveyId} (exposed as argument \code{id}),
\code{useLabels} (exposed as argument \code{use_labels}), and \code{format},
(which must be \code{json}).
}
\seealso{
\code{\link{design}}
}
