% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.r
\name{responses}
\alias{responses}
\title{Get survey responses}
\usage{
responses(id, labels = TRUE)
}
\arguments{
\item{id}{A Qualtrics survey identifier}

\item{labels}{Use question labels and answer codes (default), instead of
identifiers like "QID1"}
}
\value{
A data.table of survey responses
}
\description{
Get survey responses
}

