% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.r
\name{surveys}
\alias{surveys}
\title{Download survey metadata}
\usage{
surveys(...)
}
\arguments{
\item{...}{Further arguments to \code{\link{request}}.}
}
\value{
A data.table of survey metadata
}
\description{
Retrieve names, identifiers, and other metadata for all available surveys.
The result is sorted by the \code{lastModified} timestamp in descending
order.
}

