% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_responses.r
\name{values_to_labels}
\alias{labels_to_values}
\alias{values_to_labels}
\title{Recode survey responses}
\usage{
values_to_labels(tbl, design, keep = FALSE)

labels_to_values(tbl, design, keep = FALSE)
}
\arguments{
\item{tbl}{A data.frame of survey responses.}
}
\value{
The data.frame with recoded responses.
}
\description{
\code{values_to_labels} recodes survey responses that contain choice ids or
recodes (like \code{"5"}) to descriptive labels (like \code{"strongly
approve"}). \code{labels_to_values} does the reverse, replacing descriptive
labels with choice ids or recodes.
}

